pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vmware-storage"
version = "1.2.0"
description = "VMware vSphere storage management: datastores, iSCSI, vSAN. Domain-focused MCP skill."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Wei Zhou / 周崴" }]
dependencies = [
    "pyvmomi>=8.0.3.0,<10.0",
    "pyaml>=24.0,<27.0",
    "typer>=0.12,<1.0",
    "rich>=13.0,<15.0",
    "python-dotenv>=1.0,<2.0",
    "mcp[cli]>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10.0",
    "pytest-cov>=5.0,<8.0",
    "ruff>=0.5,<1.0",
]

[project.scripts]
vmware-storage = "vmware_storage.cli:app"
vmware-storage-mcp = "mcp_server.server:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require vCenter)",
]
